fix: 2 - DEREFERENCE_NULLABLE - long value = longAdapter.read(in).longValue(); - dereferenced expression longAdapter.read(in) is @Nullable

+++ gson/src/main/java/com/google/gson/Gson.java
@@ -453,7 +453,7 @@ public final class Gson {
         List<Long> list = new ArrayList<Long>();
         in.beginArray();
         while (in.hasNext()) {
-          long value = longAdapter.read(in).longValue();
+          Long value = longAdapter.read(in);
           list.add(value);
         }
         in.endArray();
